Google auth, protected route for create, update, delete post, one like per user

+++ src/pages/api/firebase/posts/[id]/delete.ts
@@ -1,7 +1,13 @@
-import { NextApiRequest, NextApiResponse } from "next";
+import { NextApiResponse } from "next";
 import { db } from "../../../../../../firebase-admin";
+import withAuth, {
+  CustomNextApiRequest,
+} from "../../../../../authMiddleware/withAuth";
 
-export default async (req: NextApiRequest, res: NextApiResponse) => {
+const handler: any = async (
+  req: CustomNextApiRequest,
+  res: NextApiResponse
+) => {
   if (req.method !== "DELETE")
     return res.status(400).json({ message: "BAD REQUEST" });
 
@@ -10,9 +16,13 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
   const post = await db.collection("posts").doc(id).get();
 
   if (post.exists) {
+    if (req.user.uid !== post.data()?.user.uid)
+      return res.status(401).json({ message: "UnAuthorized" });
     await post.ref.delete();
     return res.status(201).json({ message: "post deleted", data: post.id });
   } else {
     return res.status(404).json({ message: "post not found" });
   }
 };
+
+export default withAuth(handler);
